rest-api/requirements.txt
Flask>=2.1.0
pg8000==1.29.8
Flask-SQLAlchemy==3.1.1
cloud-sql-python-connector==1.2.4
gunicorn==21.0.0 
# functions-framework==3.3.0
Werkzeug==2.3.7
marshmallow-sqlalchemy==0.29.0
flask-marshmallow==0.15.0
firebase-admin==6.2.0
structlog==23.1.0
urllib3<2.0.0 #https://stackoverflow.com/questions/76175361/firebase-authentication-httpresponse-object-has-no-attribute-strict-status
Flask-Migrate==4.0.5
PyJWT==2.8.0

# websocket support 
bidict==0.22.0
cachelib==0.9.0
click==8.1.3
# Flask==2.2.5
# Flask-Login==0.6.2
# Flask-Session==0.4.0
Flask-SocketIO==5.3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.1
python-engineio
python-socketio
six==1.11.0
zipp==3.8.1
eventlet==0.33.3
